+++ contracts/governance/test/BaseTest.t.sol
@@ -23,17 +23,12 @@ import "forge-std/Test.sol";
 import "forge-std/console.sol";
 import "@openzeppelin/token/ERC20/ERC20.sol";
 import "@openzeppelin/proxy/ERC1967/ERC1967Proxy.sol";
-import "./ZRXMock.sol";
+import "./mocks/ZRXMock.sol";
 import "../src/ZRXWrappedToken.sol";
 import "../src/ZeroExVotes.sol";
 import "../src/ZeroExTimelock.sol";
 import "../src/ZeroExProtocolGovernor.sol";
 import "../src/ZeroExTreasuryGovernor.sol";
-
-function predict(address deployer, uint256 nonce) pure returns (address) {
-    require(nonce > 0 && nonce < 128);
-    return address(uint160(uint256(keccak256(abi.encodePacked(bytes2(0xd694), deployer, bytes1(uint8(nonce)))))));
-}
 
 contract BaseTest is Test {
     address payable internal account1 = payable(vm.addr(1));
@@ -54,11 +49,10 @@ contract BaseTest is Test {
         vm.deal(securityCouncil, 1e20);
     }
 
-    function setupGovernance()
-        internal
-        returns (IERC20, ZRXWrappedToken, ZeroExVotes, ZeroExTimelock, ZeroExTimelock, address, address)
-    {
-        (IERC20 zrxToken, ZRXWrappedToken token, ZeroExVotes votes) = setupZRXWrappedToken();
+    function setupGovernance(
+        IERC20 zrxToken
+    ) internal returns (ZRXWrappedToken, ZeroExVotes, ZeroExTimelock, ZeroExTimelock, address, address) {
+        (ZRXWrappedToken token, ZeroExVotes votes) = setupZRXWrappedToken(zrxToken);
 
         vm.startPrank(account1);
         address[] memory proposers = new address[](0);
@@ -86,31 +80,88 @@ contract BaseTest is Test {
         treasuryTimelock.grantRole(treasuryTimelock.CANCELLER_ROLE(), address(treasuryGovernor));
         vm.stopPrank();
 
-        return (
-            zrxToken,
-            token,
-            votes,
-            protocolTimelock,
-            treasuryTimelock,
-            address(protocolGovernor),
-            address(treasuryGovernor)
-        );
+        return (token, votes, protocolTimelock, treasuryTimelock, address(protocolGovernor), address(treasuryGovernor));
     }
 
-    function setupZRXWrappedToken() internal returns (IERC20, ZRXWrappedToken, ZeroExVotes) {
+    function setupZRXWrappedToken(IERC20 zrxToken) internal returns (ZRXWrappedToken, ZeroExVotes) {
         vm.startPrank(account1);
-        bytes memory _bytecode = vm.getCode("./ZRXToken.json");
-        IERC20 zrxToken;
-        assembly {
-            zrxToken := create(0, add(_bytecode, 0x20), mload(_bytecode))
-        }
-        address wTokenPrediction = predict(account1, vm.getNonce(account1) + 2);
+        address wTokenPrediction = predictAddress(account1, vm.getNonce(account1) + 2);
         ZeroExVotes votesImpl = new ZeroExVotes(wTokenPrediction, quadraticThreshold);
         ERC1967Proxy votesProxy = new ERC1967Proxy(address(votesImpl), abi.encodeCall(votesImpl.initialize, ()));
         ZRXWrappedToken wToken = new ZRXWrappedToken(zrxToken, ZeroExVotes(address(votesProxy)));
         vm.stopPrank();
 
         assert(address(wToken) == wTokenPrediction);
-        return (zrxToken, wToken, ZeroExVotes(address(votesProxy)));
+
+        return (wToken, ZeroExVotes(address(votesProxy)));
+    }
+
+    function mockZRXToken() internal returns (IERC20 zrxToken) {
+        vm.startPrank(account1);
+        bytes memory _bytecode = vm.getCode("./ZRXToken.json");
+        assembly {
+            zrxToken := create(0, add(_bytecode, 0x20), mload(_bytecode))
+        }
+        vm.stopPrank();
+    }
+
+    // Sourced from https://github.com/grappafinance/core/blob/master/src/test/utils/Utilities.sol
+    function predictAddress(address _origin, uint256 _nonce) public pure returns (address) {
+        if (_nonce == 0x00) {
+            return
+                address(
+                    uint160(uint256(keccak256(abi.encodePacked(bytes1(0xd6), bytes1(0x94), _origin, bytes1(0x80)))))
+                );
+        }
+        if (_nonce <= 0x7f) {
+            return
+                address(
+                    uint160(uint256(keccak256(abi.encodePacked(bytes1(0xd6), bytes1(0x94), _origin, uint8(_nonce)))))
+                );
+        }
+        if (_nonce <= 0xff) {
+            return
+                address(
+                    uint160(
+                        uint256(
+                            keccak256(
+                                abi.encodePacked(bytes1(0xd7), bytes1(0x94), _origin, bytes1(0x81), uint8(_nonce))
+                            )
+                        )
+                    )
+                );
+        }
+        if (_nonce <= 0xffff) {
+            return
+                address(
+                    uint160(
+                        uint256(
+                            keccak256(
+                                abi.encodePacked(bytes1(0xd8), bytes1(0x94), _origin, bytes1(0x82), uint16(_nonce))
+                            )
+                        )
+                    )
+                );
+        }
+        if (_nonce <= 0xffffff) {
+            return
+                address(
+                    uint160(
+                        uint256(
+                            keccak256(
+                                abi.encodePacked(bytes1(0xd9), bytes1(0x94), _origin, bytes1(0x83), uint24(_nonce))
+                            )
+                        )
+                    )
+                );
+        }
+        return
+            address(
+                uint160(
+                    uint256(
+                        keccak256(abi.encodePacked(bytes1(0xda), bytes1(0x94), _origin, bytes1(0x84), uint32(_nonce)))
+                    )
+                )
+            );
     }
 }