+++ contracts/zero-ex/contracts/src/transformers/bridges/BridgeProtocols.sol
@@ -55,5 +55,6 @@ library BridgeProtocols {
     uint128 internal constant WOOFI = 31;
     uint128 internal constant AAVEV3 = 32;
     uint128 internal constant KYBERELASTIC = 33;
-    uint128 internal constant TRADERJOEV2 = 34;
+    uint128 internal constant BARTER = 34;
+    uint128 internal constant TRADERJOEV2 = 35;
 }